refactor(chat): extract chat_manager/deps_event.rs (#292)
Task 8 of the split plan: EhDeps and its EventHandlerDeps impl move
into their own file, owning their own construction via a build()
helper the constructor now calls.
